requirements.txt
aiohttp==3.7.3
python-dotenv==0.15.0
discord.py==1.6.0